задачи повышенного уровня сложности выполнены
